+++ R/scrape_google_cmr_data.R
@@ -96,26 +96,41 @@ parse_line_graph_bitmap <- function(bitmap,
   ub = upper_bound + oset
   dplyr::tibble(
     date = days,
-    measure = round(ub - values*(ub - lb), 2)
+    measure = ub - values*(ub - lb)
   )
 }
 
-extract_clevel_line_graphs <- function(pdf) {
+extract_clevel_line_graphs <- function(pdf, date_start, date_end) {
   iso2c <- substr(basename(pdf), 12, 13)
   bitmaps <- extract_line_graph_bitmaps(pdf, pages = 1:2)
-  parse_line_graph_bitmap(bitmaps[[1]][[1]]) %>%
+  parse_line_graph_bitmap(bitmaps[[1]][[1]], date_start, date_end) %>%
     dplyr::rename(retail_recreation = .data$measure) %>%
     dplyr::mutate(iso2c = iso2c) %>%
     dplyr::select(.data$iso2c, .data$date, .data$retail_recreation) %>%
-    dplyr::left_join(parse_line_graph_bitmap(bitmaps[[1]][[2]]), by = "date") %>%
+    dplyr::left_join(
+      parse_line_graph_bitmap(bitmaps[[1]][[2]], date_start, date_end),
+      by = "date"
+    ) %>%
     dplyr::rename(grocery_pharmacy = .data$measure) %>%
-    dplyr::left_join(parse_line_graph_bitmap(bitmaps[[1]][[3]]), by = "date") %>%
+    dplyr::left_join(
+      parse_line_graph_bitmap(bitmaps[[1]][[3]], date_start, date_end),
+      by = "date"
+    ) %>%
     dplyr::rename(parks = .data$measure) %>%
-    dplyr::left_join(parse_line_graph_bitmap(bitmaps[[2]][[1]]), by = "date") %>%
+    dplyr::left_join(
+      parse_line_graph_bitmap(bitmaps[[2]][[1]], date_start, date_end),
+      by = "date"
+    ) %>%
     dplyr::rename(transit_stations = .data$measure) %>%
-    dplyr::left_join(parse_line_graph_bitmap(bitmaps[[2]][[2]]), by = "date") %>%
+    dplyr::left_join(
+      parse_line_graph_bitmap(bitmaps[[2]][[2]], date_start, date_end),
+      by = "date"
+    ) %>%
     dplyr::rename(workplaces = .data$measure) %>%
-    dplyr::left_join(parse_line_graph_bitmap(bitmaps[[2]][[3]]), by = "date") %>%
+    dplyr::left_join(
+      parse_line_graph_bitmap(bitmaps[[2]][[3]], date_start, date_end),
+      by = "date"
+    ) %>%
     dplyr::rename(residential = .data$measure)
 }
 
@@ -135,8 +150,15 @@ extract_clevel_line_graphs <- function(pdf) {
 #'     resulting data are noisy but should be accurate by +/- one day and
 #'     +/- 2 percent. Defaults to \code{TRUE}.
 #' @param pdf_dir Path to the directory where the PDF files are stored.
-#'     If \code{NULL} (the default), the code will download the PDFs from
-#'     Google's website (see above). See \code{google_download_cmr_pdfs()}
+#'     If \code{NULL} (the default), the code will download the current PDFs
+#'     from Google's website (see above). See \code{download_google_cmr_pdfs()}
+#'     for downloading the PDFs.
+#' @param use_all_dates Defaults to \code{FALSE}. If \code{TRUE}, it parses
+#'     all PDFs from various dates available in \code{pdf_dir} and averages
+#'     over the values found. This extends the time span of the data as
+#'     older Google reports contain values for earlier dates. If \code{TRUE},
+#'     make sure that \code{pdf_dir} points to the directory containing the
+#'     date indexed subdirectories. See \code{download_google_cmr_pdfs()}
 #'     for downloading the PDFs.
 #' @param silent Whether you want the function to send some status messages to
 #'     the console. Might be informative as downloading will take some time
@@ -173,7 +195,8 @@ extract_clevel_line_graphs <- function(pdf) {
 #'
 #' @export
 scrape_google_cmr_data <- function(daily_data = TRUE, pdf_dir = NULL,
-                                     silent = FALSE, cached = FALSE) {
+                                   use_all_dates = FALSE, silent = FALSE,
+                                   cached = FALSE) {
 
   parse_cmr_ctry <- function(url, parse_line_graphs = FALSE) {
 
@@ -229,7 +252,14 @@ scrape_google_cmr_data <- function(daily_data = TRUE, pdf_dir = NULL,
       return(rv)
     }
 
-    df <- extract_clevel_line_graphs(pdf)
+    date_start <- dplyr::case_when(
+      rv$date == lubridate::ymd("2020-03-29") ~ lubridate::ymd("2020-02-16"),
+      rv$date == lubridate::ymd("2020-04-05") ~ lubridate::ymd("2020-02-23"),
+      rv$date == lubridate::ymd("2020-04-11") ~ lubridate::ymd("2020-02-29")
+    )
+    date_end <- rv$date
+
+    df <- extract_clevel_line_graphs(pdf, date_start, date_end)
     if (!silent) message("done!")
     list(rv, df)
   }
@@ -247,33 +277,61 @@ scrape_google_cmr_data <- function(daily_data = TRUE, pdf_dir = NULL,
     rv <- readRDS(gzcon(url("https://raw.githubusercontent.com/joachim-gassen/tidycovid19/master/cached_data/google_cmr.RDS")))
     if (!silent) message(sprintf("done. Timestamp is %s", rv[[1]]$timestamp[1]))
   } else {
+    if (use_all_dates && is.null(pdf_dir))
+      stop("'use_all_dates' == 'TRUE' requires 'pdf_dir' to be set")
     if (!silent && is.null(pdf_dir)) message("Start downloading and scraping Google CMR data\n")
     else if (!silent) message("Start scraping Google CMR data\n")
 
     if(is.null(pdf_dir)) {
       cmr_url <- "https://www.google.com/covid19/mobility/"
 
-      urls <- xml2::read_html(cmr_url) %>%
+      urls <- list(xml2::read_html(cmr_url) %>%
         rvest::html_nodes(css = ".download-link") %>%
-        rvest::html_attr('href')
-    } else urls <- list.files(path = pdf_dir, pattern = ".pdf", full.names = TRUE)
-
-    ctry_urls <- urls[!stringr::str_detect(urls, "US_.*_Mobility")]
-
-    lst <- lapply(ctry_urls, parse_cmr_ctry, daily_data)
-    if (!daily_data) rv <- list(do.call(rbind, lst))
-    else {
-      rv <- list(
-        do.call(rbind, lapply(lst, function(x) x[[1]])),
-        do.call(rbind, lapply(lst, function(x) x[[2]]))
-      )
+        rvest::html_attr('href'))
+    } else {
+      if (use_all_dates) {
+        pdf_dir <- list.dirs(pdf_dir)
+        urls <- lapply(pdf_dir[2:length(pdf_dir)],
+                          list.files, pattern = ".pdf", full.names = TRUE)
+      }
+      else urls <- list(list.files(path = pdf_dir, pattern = ".pdf", full.names = TRUE))
     }
 
-    rv <- lapply(rv, function(x) x %>%
-      dplyr::mutate(iso3c = countrycode::countrycode(.data$iso2c, origin = "iso2c",
-                                                     destination = "iso3c")) %>%
-      dplyr::select(.data$iso3c, dplyr::everything(), -.data$iso2c) %>%
-      dplyr::mutate(timestamp = Sys.time()))
+    parse_urls <- function(urls) {
+      ctry_urls <- urls[!stringr::str_detect(urls, "US_.*_Mobility")]
+
+      lst <- lapply(ctry_urls, parse_cmr_ctry, daily_data)
+      if (!daily_data) rv <- list(do.call(rbind, lst))
+      else {
+        rv <- list(
+          do.call(rbind, lapply(lst, function(x) x[[1]])),
+          do.call(rbind, lapply(lst, function(x) x[[2]]))
+        )
+      }
+
+      rv <- lapply(rv, function(x) x %>%
+                     dplyr::mutate(iso3c = countrycode::countrycode(.data$iso2c, origin = "iso2c",
+                                                                    destination = "iso3c")) %>%
+                     dplyr::select(.data$iso3c, dplyr::everything(), -.data$iso2c) %>%
+                     dplyr::mutate(timestamp = Sys.time()))
+      rv
+    }
+
+    rlist <- lapply(urls, parse_urls)
+
+    rndmean <- function(x) {
+      if (lubridate::is.timepoint(x)) max(x)
+      else if (all(is.na(x))) NA
+      else round(mean(x, na.rm = TRUE), 2)
+    }
+
+    gcmr_c <- do.call(rbind, lapply(rlist, function(x) x[[1]])) %>%
+      dplyr::arrange(.data$iso3c, .data$date)
+    gcmr_cd <- do.call(rbind, lapply(rlist, function(x) x[[2]])) %>%
+      dplyr::group_by(.data$iso3c, .data$date) %>%
+      dplyr::summarise_at(dplyr::vars(dplyr::everything()), rndmean)
+
+    rv <- list(gcmr_c, gcmr_cd)
 
     if (!silent) message("\nDone scraping Google CMR data\n")
   }